+++ test/plugins/scenarios/trace-activation/test_service.go
@@ -97,3 +97,23 @@ func testCorrelation() {
 		return
 	}
 }
+
+func testComponent() {
+	trace.CreateLocalSpan("testComponent")
+	trace.SetComponent(5006)
+	trace.StopSpan()
+}
+
+func testAsyncInCrossGoroutine() {
+	var ch = make(chan string)
+	s, _ := trace.CreateLocalSpan("testAsyncInCrossGoroutine")
+	s.PrepareAsync()
+	trace.StopSpan()
+	go func() {
+		s.SetTag("testAsyncTag", "success")
+		s.AddLog("testAsyncLog", "success")
+		s.AsyncFinish()
+		ch <- ""
+	}()
+	<-ch
+}